+++ level_2/push_swap/src/stack-search.c
@@ -6,17 +6,17 @@
 /*   By: lde-cast <lde-cast@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/12 20:35:26 by lde-cast          #+#    #+#             */
-/*   Updated: 2023/09/02 20:46:01 by lde-cast         ###   ########.fr       */
+/*   Updated: 2023/09/08 23:34:11 by lde-cast         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stack_of.h>
 #include <stdlib.h>
 
-int	stack_pos(t_stack *head, long data)
+long	stack_pos(t_stack *head, long data)
 {
 	t_stack	*upd;
-	int		i;
+	long	i;
 
 	i = 0;
 	upd = head;
